Front end
I was able to create my branch and update the route for /sidewalk-sale. the home page is styled, with a few notes of what is missing. Routes still need to be linked. Login/reg display still needs to be set up. I'm waiting for our erd model for the users to be completed.