Contact page
Contact page is complete.
Users can fill out a form and send an email.

Right now, the email is hard coded to go to one email address
We'll need to create users and be able to pass their email into the proper field.
I'll need to share the application file with you guys so you know what to put in to get the email to work